ehr-navigator-agent/requirements.txt
flask
gunicorn
langchain-google-vertexai
langchain
langchain-core
langchain-community
google-api-python-client
requests
google-cloud-aiplatform
langgraph
pyarrow
